refactor(structure/components): refactor file structure as well as components - lots to do
