+++ addon/components/interfaces/time-picker.js
@@ -3,7 +3,7 @@
  *
  */
 import Ember from 'ember';
-import layout from '../templates/components/paper-time-picker';
+import layout from '../../templates/components/interfaces/time-picker';
 import Snap from 'snap-svg';
 import mina from 'mina';
 import moment from 'moment';
@@ -20,9 +20,9 @@ import moment from 'moment';
 //  - unit tests for any other obvious functionality that should be tested.
 
 /**
- * `Component/PaperTimePicker`
+ * `Component/TimePicker`
  *
- * @class PaperTimePicker
+ * @class TimePicker
  * @namespace Components
  * @extends Ember.Component
  */
@@ -32,9 +32,9 @@ export default Ember.Component.extend(
      * @private
      * @property classNames
      * @type String
-     * @default paper-time-picker
+     * @default time-picker
      */
-    classNames: ['paper-time-picker'],
+    classNames: ['time-picker'],
     layout: layout,
 
     /**
@@ -533,8 +533,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * converts moment object to a unix timestamp, and sets that to the global timestamp
      *
      * @private
@@ -548,8 +548,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * returns the hour of the current timestamp
      *
      * @private
@@ -565,8 +565,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * returns the minute of the current timestamp
      *
      * @private
@@ -582,8 +582,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * returns the date of the current timestamp
      *
      * @private
@@ -598,8 +598,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * sets the correct hour based on the rotated degrees on the hour drag
      *
      * @private
@@ -623,8 +623,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * sets the correct minute based on the rotated degrees on the minute drag
      *
      * @private
@@ -652,8 +652,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * returns true if the minute is a multiple of five
      *
      * @private
@@ -674,8 +674,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * returns false if the minute exceeds min or max date
      *
      * @private
@@ -715,8 +715,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * returns false if the hour exceeds min or max date
      *
      * @private
@@ -779,8 +779,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * returns true if the set timestamp is AM
      *
      * @private
@@ -801,8 +801,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * returns the minute passed in - formatted correctly
      *
      * @private
@@ -823,8 +823,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * returns the hour passed in for the header
      *
      * @private
@@ -845,8 +845,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * returns the hour passed in - formatted correctly
      *
      * @private
@@ -867,8 +867,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * formats string to an integer
      *
      * @private
@@ -883,8 +883,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * returns object with names of all hour strings
      *
      * @private
@@ -902,8 +902,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * returns object with names of all minute strings
      *
      * @private
@@ -921,8 +921,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is another good unit test function
-	 *
+   * TODO: This is another good unit test function
+   *
      * sets the timestamp to be the passed minute
      *
      * @private
@@ -940,10 +940,10 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: Anything using this snap library should
-	 * be converted to some util that gets called from within this
-	 * class.
-	 *
+   * TODO: Anything using this snap library should
+   * be converted to some util that gets called from within this
+   * class.
+   *
      * removes the hour that was passed in from the clock
      *
      * @private
@@ -962,8 +962,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: Add to snap util class
-	 *
+   * TODO: Add to snap util class
+   *
      * removes the minute that was passed in from the clock
      *
      * @private
@@ -986,8 +986,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: Add to snap util class
-	 *
+   * TODO: Add to snap util class
+   *
      * activates hour on the clock
      *
      * @private
@@ -1006,8 +1006,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: Add to snap util class
-	 *
+   * TODO: Add to snap util class
+   *
      * activates minute with text on the clock
      *
      * @private
@@ -1026,8 +1026,8 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: Add to snap util class
-	 *
+   * TODO: Add to snap util class
+   *
      * activates minute on the clock that doesn't have text
      *
      * @private
@@ -1044,10 +1044,10 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: this would be a great angle calculation util class
-	 * also this needs to have a unit test
-	 *
-	 *
+   * TODO: this would be a great angle calculation util class
+   * also this needs to have a unit test
+   *
+   *
      * gets the angle at which the drag is taking place
      *
      * @private
@@ -1068,11 +1068,11 @@ export default Ember.Component.extend(
     },
 
     /**
-	 * TODO: This is doing a lot see if you can take some of this functionality
-	 * out and move it to some kind of utils methods.
-	 * Also this needs to be unit tested.
-	 *
-	 *
+   * TODO: This is doing a lot see if you can take some of this functionality
+   * out and move it to some kind of utils methods.
+   * Also this needs to be unit tested.
+   *
+   *
      * handles all the function events for dragging on the hours clock
      * newDrag must contain start, move and stop functions within it
      *
