+++ addon/templates/components/interfaces/combined-picker.hbs
@@ -2,11 +2,11 @@
 <div class="picker-container">
 
     {{#if isClock}}
-        {{paper-time-picker timestamp=timestamp minDate=minDate maxDate=maxDate}}
+        {{interfaces/time-picker timestamp=timestamp minDate=minDate maxDate=maxDate}}
     {{/if}}
 
     {{#if isCalender}}
-        {{paper-date-picker timestamp=timestamp minDate=minDate maxDate=maxDate}}
+        {{interfaces/date-picker timestamp=timestamp minDate=minDate maxDate=maxDate}}
     {{/if}}
 
     <div class="picker-footer">
