+++ addon/templates/components/paper-datetime-picker.hbs
@@ -33,5 +33,5 @@
 </div>
 
 {{#if showInTimeClock}}
-    {{complete-time-picker timestamp=timestamp maxDate=maxDate minDate=minDate}}
+    {{interfaces/combined-picker timestamp=timestamp maxDate=maxDate minDate=minDate}}
 {{/if}}
