--- app/components/complete-time-picker.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-paper-time-picker/components/complete-time-picker';
