+++ app/components/interfaces/combined-picker.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-paper-time-picker/components/interfaces/combined-picker';
