--- app/components/paper-date-picker.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-paper-time-picker/components/paper-date-picker';
