+++ app/styles/app.scss
@@ -1,4 +1,4 @@
-@import 'time-picker/paper-time-picker';
-@import 'time-picker/complete-time-picker';
-@import 'time-picker/paper-date-picker';
+@import 'time-picker/interfaces/time-picker';
+@import 'time-picker/interfaces/combined-picker';
+@import 'time-picker/interfaces/date-picker';
 @import 'time-picker/paper-datetime-picker';
