+++ app/styles/time-picker/interfaces/combined-picker.scss
@@ -1,5 +1,5 @@
 
-.complete-time-picker {
+.combined-picker {
   // height: 330px;
   width: 250px;
   border: 1px solid black;
