+++ app/styles/time-picker/interfaces/date-picker.scss
@@ -1,4 +1,4 @@
-.paper-date-picker {
+.date-picker {
     height: 340px;
     width: 250px;
