+++ app/styles/time-picker/interfaces/time-picker.scss
@@ -1,5 +1,5 @@
 
-.paper-time-picker {
+.time-picker {
     height: 285px;
     width: 250px;
