+++ tests/integration/components/interfaces/combined-picker-test.js
@@ -2,26 +2,26 @@ import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 import moment from 'moment';
 
-moduleForComponent('complete-time-picker', 'Integration | Component | complete time picker', {
+moduleForComponent('interfaces/combined-picker', 'Integration | Component | combined picker', {
   integration: true
 });
 
 test('it renders', function(assert) {
 
-    this.set('timestamp', moment());
+    this.set('timestamp', moment().unix() * 1000);
 
-    this.render(hbs`{{complete-time-picker timestamp=timestamp}}`);
+    this.render(hbs`{{interfaces/combined-picker timestamp=timestamp}}`);
 
     assert.ok(this.$().text().trim());
 });
 
 test('toggle from clock to calender and back', function(assert) {
 
-    this.set('timestamp', moment());
-    let time = this.get('timestamp').format('hh:mm A');
-    let date = this.get('timestamp').format('MMM DD, YYYY');
+    this.set('timestamp', moment().unix() * 1000);
+    let time = moment(this.get('timestamp')).format('hh:mm A');
+    let date = moment(this.get('timestamp')).format('MMM DD, YYYY');
 
-    this.render(hbs`{{complete-time-picker timestamp=timestamp}}`);
+    this.render(hbs`{{interfaces/combined-picker timestamp=timestamp}}`);
 
     this.$('.current-date').click();
     assert.equal(this.$('.current-date').text(), time);
