+++ tests/unit/components/interfaces/combined-picker-test.js
@@ -1,0 +1,16 @@
+import { moduleForComponent, test } from 'ember-qunit';
+
+moduleForComponent('interfaces/combined-picker', 'Unit | Component | interfaces/combined picker', {
+  // Specify the other units that are required for this test
+  // needs: ['component:foo', 'helper:bar'],
+  unit: true
+});
+
+test('it renders', function(assert) {
+  
+  // Creates the component instance
+  /*let component =*/ this.subject();
+  // Renders the component to the page
+  this.render();
+  assert.equal(this.$().text().trim(), '');
+});
