+++ tests/unit/components/interfaces/time-picker-test.js
@@ -1,0 +1,16 @@
+import { moduleForComponent, test } from 'ember-qunit';
+
+moduleForComponent('interfaces/time-picker', 'Unit | Component | interfaces/time picker', {
+  // Specify the other units that are required for this test
+  // needs: ['component:foo', 'helper:bar'],
+  unit: true
+});
+
+test('it renders', function(assert) {
+  
+  // Creates the component instance
+  /*let component =*/ this.subject();
+  // Renders the component to the page
+  this.render();
+  assert.equal(this.$().text().trim(), '');
+});
